Hoist GraphicLine out of Map.py
